Support XDG-conformant locations for the nix channels list
Ref: https://github.com/NixOS/nix/pull/5588

+++ nixos/modules/services/misc/nix-daemon.nix
@@ -636,7 +636,7 @@ in
         install -m 0755 -d /nix/var/nix/{gcroots,profiles}/per-user
 
         # Subscribe the root user to the NixOS channel by default.
-        if [ ! -e "/root/.nix-channels" ]; then
+        if [ ! -e "/root/.nix-channels" ] && [ ! -e /root/.local/share/nix/channels ]; then
             echo "${config.system.defaultChannel} nixos" > "/root/.nix-channels"
         fi
       '';
